+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the "dhall-lang" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2]
+- Fixed URLs, Files and environment variables
+- Highlighting upper case labels as types
+
 ## [0.0.1]
 - Initial release. Basic highlighting support. Some scopes a likely wrong.